+++ Redux/redux-note/src/components/NewNote.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import { useDispatch } from 'react-redux'
+
+import { createNote } from '../reducers/noteReducer'
+
+const NewNote = () => {
+    const dispatch = useDispatch()
+
+    const addNote = (event) => {
+        event.preventDefault()
+        const content = event.target.note.value
+        event.target.note.value = ''
+        dispatch(createNote(content)) 
+      }
+
+  return (
+    <form onSubmit={addNote}>
+        <input name='note'/>
+        <button type='submit'>add</button>
+      </form>
+  )
+}
+
+export default NewNote
