+++ Redux/redux-note/src/components/Notes.jsx
@@ -1,0 +1,30 @@
+import React from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { toggleImportanceOf } from '../reducers/noteReducer'
+
+const Note = ({note, handleClick}) => {
+  return (
+    <li onClick={handleClick}>
+        {note.content} <strong>{note.important ? 'important' : ''}</strong>
+    </li>
+  )
+}
+
+const Notes = () => {
+    const dispatch = useDispatch()
+    const notes = useSelector(state => state)
+    console.log(notes)
+    return (
+        <ul>
+            {notes.map(note =>
+            <Note
+                key = {note.id}
+                note = {note}
+                handleClick = {() => dispatch(toggleImportanceOf(note.id))}
+            />
+            )}
+        </ul>
+    )
+  }
+
+export default Notes
